[benchmark] refactor killing of providers in run performance script
* only kill if provider has been started before
* use pkill only in case where maven is used

+++ tests/performance-test/run-performance-tests.sh
@@ -333,10 +333,18 @@ function stopAnyProvider {
     # pkill is required if maven is used to start a provider. Maven launches the
     # provider as a child process, which seems not to be killed automatically along
     # with the parent process
-    pkill -P $PROVIDER_PID
-    kill $PROVIDER_PID
-    wait $PROVIDER_PID
-    PROVIDER_PID=""
+    if [ "$PROVIDER_PID" != "" ]
+    then
+        if [ "$USE_MAVEN" != "ON" ]
+        then
+            echo "do not call pkill for provider id $PROVIDER_ID"
+        else
+            pkill -P $PROVIDER_PID
+        fi
+        kill $PROVIDER_PID
+        wait $PROVIDER_PID
+        PROVIDER_PID=""
+    fi
 }
 
 function echoUsage {
